# api/openapi.yaml
openapi: 3.1.0
info:
  title: portal API
  version: 1.0.0
  description: |
    The portal HTTP API. Every route chi serves under /api/v1
    (internal/server/server.go) is described here; the schemas mirror the
    handler-layer *Response types in internal/handler (the wire boundary —
    repository rows never serialize directly).

    web/src/api/types.ts is generated from this file with openapi-typescript
    (`npm run generate:api` in web/); openapi-fetch consumes the generated
    `paths` interface. Edit this spec, regenerate, and the frontend contract
    follows — CI fails on drift.

    Two surfaces sit outside the JSON request/response model:
    - GET /workspaces/{workspaceId}/runs/{runId}/logs/ws upgrades to a
      WebSocket (documented on the operation; the stream itself is
      newline-delimited log text, authenticated via the
      `bearer` + token subprotocol pair).
    - GET /auth/github and /auth/github/callback are browser redirect hops of
      the OAuth flow, not XHR endpoints.

    Errors use one envelope everywhere (ErrorResponse) via respond.Error /
    respond.FromError; operations declare it as their default response.
servers:
  - url: /api/v1
paths:
...
  /auth/github:
    get:
      operationId: githubLogin
      summary: Start the GitHub OAuth flow (browser redirect)
      security: []
      responses:
        "307":
          description: Redirect to GitHub's authorize URL; sets the signed oauth_state cookie.
        default: { $ref: "#/components/responses/Error" }
  /auth/github/callback:
    get:
      operationId: githubCallback
      summary: GitHub OAuth callback (browser redirect)
      description: >
        Exchanges the code, provisions the user, sets the short-lived HttpOnly
        handoff cookie (scoped to /auth/handoff), and redirects to the SPA
        callback page, which completes login via POST /auth/handoff.
      security: []
      parameters:
        - name: code
          in: query
          schema: { type: string }
        - name: state
          in: query
          schema: { type: string }
      responses:
        "307":
          description: Redirect to the SPA /auth/callback page.
        default: { $ref: "#/components/responses/Error" }
  /auth/handoff:
    post:
      operationId: authHandoff
      summary: Exchange the HttpOnly login-handoff cookie for the JWT (one-time)
      description: >
        Completes the login handoff. The OAuth callback (and dev login) set a
        short-lived HttpOnly cookie scoped to this path; this endpoint returns
        the JWT once and expires the cookie in the same response
        (delete-on-read), so a second call gets 401. POST keeps the exchange
        non-cacheable, and SameSite=Lax blocks cross-site POSTs from carrying
        the cookie. Unauthenticated by design — the caller has no bearer token
        yet.
      security: []
      responses:
        "200":
          description: The JWT, exactly once; the handoff cookie is expired in the same response.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                required: [token]
        "401":
          description: Missing, expired, or malformed handoff cookie
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        default: { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: The standard error envelope (respond.Error / respond.FromError).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: string, description: "HTTP status text." }
        message: { type: string }
        request_id: { type: string, description: "Set on 5xx responses so logs can be correlated." }
      required: [error]
...
security:
  - bearerAuth: []